Adjust script to use a local cache for quicker environment switching
